// next-js/eventverse/evwj-old/types/venue.ts
export interface VendorBoothAssignment {
	invitationId: string;
	vendorId?: string;
	vendorEmail: string;
	vendorBusinessName: string;
	boothTypeId: string;
	boothTypeName: string;
	contactEmail: string;
	contactPhone?: string;
}
// ...
export interface VenueObject {
	id: number;
	name: string;
	type:
	| "stage"
	| "podium"
	| "exit"
	| "dancefloor"
	| "tent"
	| "booth"
	| "plant"
	| "pillar"
	| "fountain"
	| "lounge"
	| "bar"
	| "buffet"
	| "dessert"
	| "restroom"
	| "bathroom"
	| "coatcheck"
	| "kitchen"
	| "gifttable"
	| "guestbook"
	| "photowall"
	| "dj"
	| "emergency-exit";
	x: number;
	y: number;
	width: number;
	height: number;
	rotation?: number;
	color?: string;
	assignedVendor?: VendorBoothAssignment;
	boothTypeId?: string;
}